+++ templates/workspaces/base/terraform/workspace.tf
@@ -53,8 +53,36 @@ module "aad" {
 }
 
 module "airlock" {
-  count                  = var.enable_airlock ? 1 : 0
-  source                 = "./airlock"
+  count                                  = var.enable_airlock && var.airlock_version == 1 ? 1 : 0
+  source                                 = "./airlock"
+  location                               = var.location
+  tre_id                                 = var.tre_id
+  tre_workspace_tags                     = local.tre_workspace_tags
+  ws_resource_group_name                 = azurerm_resource_group.ws.name
+  enable_local_debugging                 = var.enable_local_debugging
+  services_subnet_id                     = module.network.services_subnet_id
+  short_workspace_id                     = local.short_workspace_id
+  airlock_processor_subnet_id            = module.network.airlock_processor_subnet_id
+  arm_environment                        = var.arm_environment
+  enable_cmk_encryption                  = var.enable_cmk_encryption
+  encryption_key_versionless_id          = var.enable_cmk_encryption ? azurerm_key_vault_key.encryption_key[0].versionless_id : null
+  encryption_identity_id                 = var.enable_cmk_encryption ? azurerm_user_assigned_identity.encryption_identity[0].id : null
+  enable_airlock_malware_scanning        = var.enable_airlock_malware_scanning
+  airlock_malware_scan_result_topic_name = var.enable_airlock_malware_scanning ? var.airlock_malware_scan_result_topic_name : null
+
+  providers = {
+    azurerm      = azurerm
+    azurerm.core = azurerm.core
+  }
+
+  depends_on = [
+    module.network,
+  ]
+}
+
+module "airlock_v2" {
+  count                  = var.enable_airlock && var.airlock_version >= 2 ? 1 : 0
+  source                 = "./airlock_v2"
   location               = var.location
   tre_id                 = var.tre_id
   tre_workspace_tags     = local.tre_workspace_tags
@@ -96,7 +124,8 @@ module "azure_monitor" {
 
   depends_on = [
     module.network,
-    module.airlock
+    module.airlock,
+    module.airlock_v2
   ]
 }